requirements.txt
azure-storage-blob==12.20.0
dash==2.17.0
dash_bootstrap_components==1.6.0
geopandas==0.14.4
gunicorn==22.0.0
pyarrow==16.1.0
